+++ src/rsa/SecretKey.java
@@ -1,17 +1,17 @@
 package rsa;
 
 public class SecretKey {
-    long d;
-    long p;
-    long q;
+    int d;
+    int p;
+    int q;
 
-    public SecretKey(long d, long p, long q) {
+    public SecretKey(int d, int p, int q) {
         this.d = d;
         this.p = p;
         this.q = q;
     }
 
-    public long fn(long p, long q) {
+    public int fn(int p, int q) {
         return (p - 1) * (q - 1);
     }
 }
